+++ src/components/Map.tsx
@@ -1,6 +1,6 @@
 import "leaflet/dist/leaflet.css";
 
-import { Icon, LeafletEventHandlerFnMap, Map as LeafletMap } from "leaflet";
+import { LeafletEventHandlerFnMap, Map as LeafletMap } from "leaflet";
 
 import {
   MapContainer,
@@ -11,27 +11,6 @@ import {
   ZoomControl,
 } from "react-leaflet";
 import { CSSProperties, PropsWithChildren, Ref, RefObject } from "react";
-
-import mapMarker from "../images/map-marker.svg";
-import mapMarkerActive from "../images/map-marker-active.svg";
-import { CareProviderSearchResult } from "../types";
-import { getResultBounds } from "../utils";
-
-const markerIcon = new Icon({ iconUrl: mapMarker, iconSize: [32, 32] });
-const markerActiveIcon = new Icon({
-  iconUrl: mapMarkerActive,
-  iconSize: [32, 32],
-});
-
-const rerenderMap = (
-  mapRef: RefObject<LeafletMap>,
-  results: CareProviderSearchResult[]
-) => {
-  setTimeout(() => {
-    mapRef.current?.invalidateSize();
-    mapRef.current?.fitBounds(getResultBounds(results), { animate: false });
-  }, 100);
-};
 
 function MapEvents({
   eventHandlers,
@@ -79,4 +58,3 @@ function Map({
 }
 
 export default Map;
-export { markerIcon, markerActiveIcon, rerenderMap };
