+++ src/pages/Search/MobileResults.tsx
@@ -4,13 +4,12 @@ import { Map as LeafletMap } from "leaflet";
 import { useState, useRef, useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { logEvent, AnalyticsAction } from "../../analytics";
-import { rerenderMap } from "../../components/Map";
 import MobileViewToggle from "../../components/Search/MobileViewToggle";
 import ResultCard from "../../components/Search/ResultCard";
 import ResultsList from "../../components/Search/ResultsList";
 import ResultsMap from "../../components/Search/ResultsMap";
 import { CareProviderSearchResult } from "../../types";
-import { getMapMarker, getResultBounds } from "../../utils";
+import { getMapMarker, getResultBounds, rerenderMap } from "../../utils";
 import { ReactComponent as Close } from "../../images/close.svg";
 
 /**
